+++ src/utils/simple_http_server.py
@@ -12,7 +12,8 @@ the Free Software Foundation, either version 3 of the License, or
 
 For more see the file 'readme/COPYING' for copying permission.
 """
-
+import re
+import sys
 import errno
 import thread
 import socket
@@ -23,6 +24,62 @@ from src.utils import settings
 from socket import error as socket_error
 from src.thirdparty.colorama import Fore, Back, Style, init
 from BaseHTTPServer import BaseHTTPRequestHandler,HTTPServer
+
+"""
+Validates IPv4 addresses.
+"""
+def is_valid_ipv4(ip_addr):
+    pattern = re.compile(r"""
+        ^
+        (?:
+          # Dotted variants:
+          (?:
+            # Decimal 1-255 (no leading 0's)
+            [3-9]\d?|2(?:5[0-5]|[0-4]?\d)?|1\d{0,2}
+          |
+            0x0*[0-9a-f]{1,2}  # Hexadecimal 0x0 - 0xFF (possible leading 0's)
+          |
+            0+[1-3]?[0-7]{0,2} # Octal 0 - 0377 (possible leading 0's)
+          )
+          (?:                  # Repeat 0-3 times, separated by a dot
+            \.
+            (?:
+              [3-9]\d?|2(?:5[0-5]|[0-4]?\d)?|1\d{0,2}
+            |
+              0x0*[0-9a-f]{1,2}
+            |
+              0+[1-3]?[0-7]{0,2}
+            )
+          ){0,3}
+        |
+          0x0*[0-9a-f]{1,8}    # Hexadecimal notation, 0x0 - 0xffffffff
+        |
+          0+[0-3]?[0-7]{0,10}  # Octal notation, 0 - 037777777777
+        |
+          # Decimal notation, 1-4294967295:
+          429496729[0-5]|42949672[0-8]\d|4294967[01]\d\d|429496[0-6]\d{3}|
+          42949[0-5]\d{4}|4294[0-8]\d{5}|429[0-3]\d{6}|42[0-8]\d{7}|
+          4[01]\d{8}|[1-3]\d{0,9}|[4-9]\d{0,8}
+        )
+        $
+    """, re.VERBOSE | re.IGNORECASE)
+    return pattern.match(ip_addr) is not None
+
+def grab_ip_addr():
+  try:
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    s.connect(("8.8.8.8",53))
+    s.settimeout(2)
+    ip_addr = s.getsockname()[0]
+    s.close()
+    return ip_addr
+  except socket_error, err_msg:
+    if errno.ECONNREFUSED:
+      warn_msg = "Internet seems unreachable."
+      print settings.print_warning_msg(warn_msg)
+    else:
+      print settings.print_critical_msg(str(err_msg)) + "\n"
+      sys.exit(0)
 
 class Handler(BaseHTTPRequestHandler):
     def do_GET(self):